Refactor HomePage component: extract ProfileSummary and TechStack into separate components, update profile summary data structure
