+++ public/assets/data/homePageData.json
@@ -1,7 +1,5 @@
 {
   "profileSummary": {
-    "name": "Javid Ahmed",
-    "title": "Software Developer",
     "description": [
       "I have a passion for learning and problem solving and I enjoy developing my skills in new technologies and methods. Since joining Capgemini Engineering, I have gained project experience of full stack software development and worked in multiple project teams. My work has included creating automated CI/CD pipelines for Python applications. I have also worked on databases and the APIs being consumed by these applications. Additionally, I have developed User Interfaces using different frameworks to create web, desktop, and mobile applications. This has involved learning to use a wide range of technologies, as well as the Agile methodologies used for these projects.",
       "Prior to starting at Capgemini Engineering, I studied Physics at the University of Birmingham and during this time, I worked on multiple scientific computing projects in Python. These involved developing and training convolutional neural networks, image processing, 2D simulations, and data analysis in a scientific context. Additionally, I developed my capacity for quick learning, writing technical reports, delivering presentations, and working effectively within a team.",
