+++ src/components/pages/home/index.ts
@@ -1,1 +1,3 @@
+export { default as ProfileSummary } from "./ProfileSummary";
 export { default as TechnologyButton } from "./TechnologyButton";
+export { default as TechStack } from "./TechStack";
